scripts: common: fix pread_full() for windows
pread() is not available for windows, convert to lseek() + read()

Signed-off-by: Maud Spierings <maudspierings@gocontroll.com>

---

I think this modifies the state of fd, I am not sure if pread() did this
too. May be necessary to point the file pointer where it was before? On
the other hand this reads the whole file so why.